% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/download.R
\name{download_voting_area_csv}
\alias{download_voting_area_csv}
\title{Download the voting area data file. On 2015-03-05 the data was not in
standard CSV form.}
\usage{
download_voting_area_csv(filename,
  url = "https://www.avoindata.fi/data/fi/dataset/df2d00d2-4a00-46ba-b16c-8f73c5e6bb32/resource/b9e48d8e-f8c3-4dde-b221-45e495021c9c/download/rvaalit2015eduskuntavaalitdataavoindatarvaalit2015eduskuntavaalitdataavoindata1003796rakaanestysalu")
}
\description{
Download the voting area data file. On 2015-03-05 the data was not in
standard CSV form.
}

