% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{find_conflicting_coordinates}
\alias{find_conflicting_coordinates}
\title{Find rows with duplicated coordinates but differing municipality or voting
area.}
\usage{
find_conflicting_coordinates(df)
}
\description{
Find rows with duplicated coordinates but differing municipality or voting
area.
}

