% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{get_central_points}
\alias{get_central_points}
\title{Get a central point for each voting area in each municipality.}
\usage{
get_central_points(df, func = median)
}
\description{
Get a central point for each voting area in each municipality.
}

