% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{load_and_clean}
\alias{load_and_clean}
\title{Download, parse, clean and return the voting area data.}
\usage{
load_and_clean()
}
\description{
Download, parse, clean and return the voting area data.
}

