% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{log_and_clean}
\alias{log_and_clean}
\title{Log the dirty rows and return the clean rows.}
\usage{
log_and_clean(df, filename, finder, ...)
}
\arguments{
\item{df}{The data frame to process.}

\item{filename}{The filename to write the log into.}

\item{finder}{The function that finds the dirty rows and returns a logical
vector.}
}
\description{
Log the dirty rows and return the clean rows.
}

