% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_voting_area_csv}
\alias{parse_voting_area_csv}
\title{Parse the voting area CSV file into a data frame.}
\usage{
parse_voting_area_csv(filename)
}
\description{
Parse the voting area CSV file into a data frame.
}

