% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clean.R
\name{separate_rows}
\alias{separate_rows}
\title{Separate a data frame into two, dirty and clean.}
\usage{
separate_rows(df, dirty_row_index)
}
\description{
Create a list of two data frames, 'dirty_df' with the rows indicated by
dirty_row_index, 'clean_df' with the rest of the rows.
}

