% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main.R
\name{transform_etrs_tm35fin_to_wgs84}
\alias{transform_etrs_tm35fin_to_wgs84}
\title{Transform ETRS-TM35FIN into WGS 84.}
\usage{
transform_etrs_tm35fin_to_wgs84(df)
}
\description{
Transform ETRS-TM35FIN into WGS 84.
}

